compile_java.py: preserve temp_dir for javac rsp file.
when --print-javac-command-line is given, javac command line
contains rspfile path as @rspfile, but rspfile was removed
at the end of compile_java.py, so ide_query can't get the
full command line.

preserve temp_dir if --print-javac-command-line is given.
temp_dir will be removed by ide_query (or `siso query ideanalysis`)

Bug: b/337982520
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5905024
Commit-Queue: Fumitoshi Ukai <ukai@google.com>
Auto-Submit: Fumitoshi Ukai <ukai@google.com>
Cr-Commit-Position: refs/heads/main@{#1364773}

+++ build/android/gyp/compile_java.py
@@ -600,7 +600,9 @@ def _RunCompiler(changes,
 
     logging.info('Completed all steps in _RunCompiler')
   finally:
-    shutil.rmtree(temp_dir)
+    # preserve temp_dir for rsp fie when --print-javac-command-line
+    if not options.print_javac_command_line:
+      shutil.rmtree(temp_dir)
 
 
 def _ParseOptions(argv):
